integrated dir walker into python script

+++ UltraFastDuplicateFilesFinder.py
@@ -1,30 +1,16 @@
 #!/usr/bin/python
-# -*- coding: utf-8 -*-
-#
-
 """
 Ultra Fast Duplicate Files Finder
-=================================
-  by Gautier Portet <kassoulet gmail com>
+
+Deletes multiple files within a directory
+
+Example use:
+
+  ./UltraFastDuplicateFilesFinder.py - dir
 
 
-Takes a list of file from stdin.
-And print the duplicate ones.
-
-
-example use:
-
-  find ~/ -size +10M | ./UltraFastDuplicateFilesFinder.py
-
-to find duplicates in your home folder, all files more than 10MB.
-
-UltraFastDuplicateFilesFinder compares only the very beginning of the files.
-Its sufficient for most uses, but use with caution.
-
-But this way is quite useful to detect duplicates within corrupted media files...
-
-
-this is public domain.
+To be fast the files are grouped by size first. Only if there are files
+with the same size a md5 checksum will be performed.
 """
 
 import sys
@@ -32,18 +18,16 @@ import os
 import hashlib
 import itertools
 
-# read one CHUNK_SIZE bytes to check duplicates
-CHUNK_SIZE = 1024
-
-# buffer size when doing whole file md5
-BUFFER_SIZE = 64*1024
-
-
 class FileInfo:
     filename = ""
     size = 0
     hash = ""
 
+# read one CHUNK_SIZE bytes to check duplicates
+CHUNK_SIZE = 1024
+
+# buffer size when doing whole file md5
+BUFFER_SIZE = 64*1024
 def get_file_hash(filename, limit_size=None, buffer_size=BUFFER_SIZE):
     """
     Return the md5 hash of given file as an hexadecimal string.
@@ -74,16 +58,6 @@ def get_file_hash(filename, limit_size=None, buffer_size=BUFFER_SIZE):
     f.close()
     return hasher.hexdigest()
 
-delpaths = ''
-if len(sys.argv) > 1 and sys.argv[1] == '-d':
-    delpaths = sys.argv[2].split(';')
-    print(delpaths)
-
-files = {}
-hashlist = {}
-totalsize = 0
-totalfiles = 0
-
 def humanize_size(size):
     """
     Return the file size as a nice, readable string.
@@ -93,42 +67,79 @@ def humanize_size(size):
         if hsize > 0.5:
             return '%.2f %s' % (hsize, suffix)
 
+def parseArgs(argv):
+    import argparse
+    parser = argparse.ArgumentParser(description='Find duplicate files.')
+    parser.add_argument('-d', metavar='dir', action='append')
+    parser.add_argument('--dry', action='store_true')
+    parser.add_argument('--min-size', type=int)
+    return parser.parse_args(argv)
+
+args = parseArgs(sys.argv[1:])
+delpaths = args.d
+dryRun = args.dry
+min_size = args.min_size
+
+files = {}
+totalsize = 0
+totalfiles = 0
+
 # we start here by checking all filesizes
-for filename in sys.stdin:
-    filename = filename.strip()
-    if not os.path.isfile(filename):
-        continue
+for path, directories, filelist in os.walk('.'):
+    for relFileName in filelist:
+        filename = os.path.join(path, relFileName)
+
+        if not os.path.isfile(filename):
+            continue
+            
+        size = os.path.getsize(filename)
+        if size < min_size:
+            continue
         
-    size = os.path.getsize(filename)
-    
-    fi = FileInfo()
-    fi.filename = filename
-    fi.size = size
-    files[filename] = fi
-    totalfiles += 1
-    totalsize += size
-    sys.stdout.write('%d files (%s)           \r' % (totalfiles, humanize_size(totalsize)))
+        fi = FileInfo()
+        fi.filename = filename
+        fi.size = size
+        files[filename] = fi
+        totalfiles += 1
+        totalsize += size
+        sys.stdout.write('%d files (%s)           \r' % (totalfiles, humanize_size(totalsize)))
 
 print ''
-# print the report
-print '%10s   %s' % ('size', 'filename')
+print ("group by size")
 
 # group files by size
-selFsize = lambda fi:fi.size
-filenamesBySize = sorted(files.values(), key=selFsize)
-for k, g in itertools.groupby(filenamesBySize, selFsize):
-    filesOfThisSize = list(g)
+hashlist = {}
+filesBySize = {}
+sizeToHash = 0
+for f in files.values():
+    if not filesBySize.has_key(f.size):
+        filesBySize[f.size] = []
+    filesBySize[f.size].append(f)
+    lGroup = len(filesBySize[f.size])
+    if lGroup == 2:
+        sizeToHash += 2*f.size
+    if lGroup > 2:
+        sizeToHash += f.size
+
+print ('calculate hashes of ' + humanize_size(sizeToHash))
+sizeHashed = 0.0
+for size, filesOfThisSize in filesBySize.iteritems():
     if len(filesOfThisSize) <= 1:
         continue
-        
+
     # calculate the hashes for this group
-    for fn in filesOfThisSize:     
+    for fn in filesOfThisSize:
         h = get_file_hash(fn.filename)
         if not hashlist.has_key(h):
         	hashlist[h] = []
         
         fn.hash=h
         hashlist[h].append(fn)
+        sizeHashed += fn.size
+        sys.stdout.write('{0:.2%}                    \r'.format(sizeHashed / sizeToHash))
+
+# print the report
+print '%10s   %s' % ('size', 'filename')
 
 nDupGroups = 0
 nDupFiles = 0
@@ -155,9 +166,14 @@ for hl,fileinfos in sorted(hashlist.iteritems(), key=lambda(k,v):v[0].size):
             filesToDelete.remove(keep)
 
         for toDel in filesToDelete:
-            os.remove(toDel.filename)
+            if not dryRun:
+                os.remove(toDel.filename)
+                print ('deleted %s' % toDel.filename)
+            else:
+                print ('simulate deletion of %s' % toDel.filename)
+            
             deletedFileSize += toDel.size
-            print ('deleted %s' % toDel.filename)
+            
 
 # final summary
 print 20*'-'
